Move all mouse_-functions to mousemanager.h
Move all the mouse related functions to mousemanager.h. From there, we
can make them member functions of MouseManager.
