+++ src/client.cpp
@@ -15,7 +15,7 @@
 #include "keymanager.h"
 #include "layout.h"
 #include "monitor.h"
-#include "mouse.h"
+#include "mousemanager.h"
 #include "root.h"
 #include "settings.h"
 #include "stack.h"
