+++ src/mouse.h
@@ -35,19 +35,6 @@ public:
 
 int mouse_binding_equals(const MouseBinding* a, const MouseBinding* b);
 
-int mouse_unbind_all();
-std::experimental::optional<MouseBinding> mouse_binding_find(unsigned int modifiers, unsigned int button);
-
-unsigned int string2button(const char* name);
-MouseFunction string2mousefunction(const char* name);
-
-void grab_client_buttons(Client* client, bool focused);
-
-void mouse_handle_event(XEvent* ev);
-void mouse_initiate_drag(Client* client, MouseDragFunction function);
-void mouse_stop_drag();
-bool mouse_is_dragging();
-void handle_motion_event(XEvent* ev);
 
 // get the vector to snap a client to it's neighbour
 void client_snap_vector(Client* client, Monitor* monitor,
@@ -55,15 +42,6 @@ void client_snap_vector(Client* client, Monitor* monitor,
                         int* return_dx, int* return_dy);
 
 bool is_point_between(int point, int left, int right);
-
-void mouse_initiate_move(Client* client, const std::vector<std::string> &cmd);
-void mouse_initiate_zoom(Client* client, const std::vector<std::string> &cmd);
-void mouse_initiate_resize(Client* client, const std::vector<std::string> &cmd);
-void mouse_call_command(Client* client, const std::vector<std::string> &cmd);
-/* some mouse drag functions */
-void mouse_function_move(XMotionEvent* me);
-void mouse_function_resize(XMotionEvent* me);
-void mouse_function_zoom(XMotionEvent* me);
 
 #endif
