+++ src/mousemanager.h
@@ -23,3 +23,26 @@ public:
 private:
     Cursor cursor;
 };
+
+int mouse_unbind_all();
+std::experimental::optional<MouseBinding> mouse_binding_find(unsigned int modifiers, unsigned int button);
+
+unsigned int string2button(const char* name);
+MouseFunction string2mousefunction(const char* name);
+
+void grab_client_buttons(Client* client, bool focused);
+
+void mouse_handle_event(XEvent* ev);
+void mouse_initiate_drag(Client* client, MouseDragFunction function);
+void mouse_stop_drag();
+bool mouse_is_dragging();
+void handle_motion_event(XEvent* ev);
+
+void mouse_initiate_move(Client* client, const std::vector<std::string> &cmd);
+void mouse_initiate_zoom(Client* client, const std::vector<std::string> &cmd);
+void mouse_initiate_resize(Client* client, const std::vector<std::string> &cmd);
+void mouse_call_command(Client* client, const std::vector<std::string> &cmd);
+/* some mouse drag functions */
+void mouse_function_move(XMotionEvent* me);
+void mouse_function_resize(XMotionEvent* me);
+void mouse_function_zoom(XMotionEvent* me);
